requirements.txt
python-telegram-bot
gspread
oauth2client
google-api-python-client
python-dotenv
telegram